+++ fuzzywuzzy/src/main/java/me/xdrop/fuzzywuzzy/FuzzyWuzzy.java
@@ -1,4 +1,4 @@
-package me.xdrop.levenshtein;
+package me.xdrop.fuzzywuzzy;
 
 public class FuzzyWuzzy {
     public static <A extends Algorithm> A algorithm(Class<A> ofClass) {
